+++ src/mechanism/tests/exists.rs
@@ -3,15 +3,15 @@
 mod capture {
 
     use crate::helpers;
+    use crate::mechanism::Kind;
     use crate::mechanism::Mechanism;
-    use crate::mechanism::MechanismKind;
 
     #[test]
     fn test_match_on_exists() {
         let string = "exists:a.example.com";
         let option_test: Option<Mechanism<String>>;
 
-        option_test = helpers::capture_matches(&string, MechanismKind::Exists);
+        option_test = helpers::capture_matches(&string, Kind::Exists);
 
         let test = option_test.unwrap();
         assert_eq!(test.is_pass(), true);